backend/server/TextractQueries/requirements.txt
amazon-textract-caller==0.0.29
amazon-textract-helper==0.0.34
amazon-textract-overlayer==0.0.11
amazon-textract-prettyprinter==0.1.2
amazon-textract-response-parser==0.1.45
apturl==0.5.2
asttokens==2.2.1
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
boto3==1.26.129
botocore==1.29.129
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
ci-info==0.3.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.8
configparser==5.3.0
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
dnspython==2.3.0
duplicity==0.8.21
etelemetry==0.3.0
executing==1.2.0
fasteners==0.14.1
filelock==3.12.0
fillpdf==0.7.2
fitz==0.0.1.dev2
future==0.18.2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipython==8.13.2
isodate==0.6.1
jedi==0.18.2
jeepney==0.7.1
jmespath==1.0.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
looseversion==1.1.2
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
marshmallow==3.19.0
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
numpy==1.24.3
oauthlib==3.2.0
olefile==0.46
packaging==23.1
pandas==2.0.1
paramiko==2.9.3
parso==0.8.3
pathlib==1.0.1
pdf2image==1.16.3
pdfrw==0.4
pdfrw2==0.5.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
protobuf==3.12.4
prov==2.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycups==2.0.1
pydot==1.4.2
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.3.3
PyMuPDF==1.22.0
PyNaCl==1.5.0
pyparsing==2.4.7
pypdf==3.9.0
PyPDF2==3.0.1
pyRFC3339==1.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==1.0.0
pytz==2022.1
pyxdg==0.27
pyxnat==1.5
PyYAML==5.4.1
rdflib==6.3.2
reportlab==3.6.8
requests==2.25.1
s3transfer==0.6.1
scipy==1.10.1
SecretStorage==3.3.1
simplejson==3.19.1
six==1.16.0
stack-data==0.6.2
systemd-python==234
tabulate==0.9.0
textract-trp==0.1.3
traitlets==5.9.0
traits==6.3.2
trp==0.1.6a1
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.6
xdg==5
xkit==0.0.0
zipp==1.0.0
